moving useBlockTime hook to the api module
