+++ src/screen/ReferendaScreen.tsx
@@ -6,7 +6,7 @@ import {ProposalInfo} from 'presentational/ProposalInfo';
 import SafeView, {noTopEdges} from 'presentational/SafeView';
 import * as React from 'react';
 import {FlatList, StyleSheet} from 'react-native';
-import {useBlockTime} from 'src/hook/useBlockTime';
+import {useBlockTime} from 'src/api/hooks/useBlockTime';
 import {useReferendums} from 'src/hook/useReferendums';
 import {useBestNumber} from 'src/hook/useVotingStatus';
 import globalStyles, {standardPadding} from 'src/styles';
